UI: refactor snapshots list routes
